extracted fetchMeasurementForFixed method and clean up
