+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveRecyclerAdapter.kt
@@ -3,6 +3,7 @@ package io.lunarlogic.aircasting.screens.dashboard.active
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.FragmentManager
+import io.lunarlogic.aircasting.models.Session
 import io.lunarlogic.aircasting.screens.dashboard.SessionsRecyclerAdapter
 
 class MobileActiveRecyclerAdapter(
@@ -20,5 +21,9 @@ class MobileActiveRecyclerAdapter(
             )
         viewMvc.registerListener(mListener)
         return MyViewHolder(viewMvc)
+    }
+
+    override fun fetchMeasurementsForFixed(session: Session) {
+        // do nothing
     }
 }
