+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/fixed/FixedRecyclerAdapter.kt
@@ -3,6 +3,8 @@ package io.lunarlogic.aircasting.screens.dashboard.fixed
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.FragmentManager
+import io.lunarlogic.aircasting.database.DatabaseProvider
+import io.lunarlogic.aircasting.models.Session
 import io.lunarlogic.aircasting.screens.dashboard.SessionsRecyclerAdapter
 
 
@@ -21,5 +23,26 @@ class FixedRecyclerAdapter(
             )
         viewMvc.registerListener(mListener)
         return MyViewHolder(viewMvc)
+    }
+
+    override fun fetchMeasurementsForFixed(session: Session) {
+        downloadMeasurementsForSession(session)
+    }
+
+    fun downloadMeasurementsForSession(session: Session) {
+        reloadSessionMeasurements(session)
+    }
+
+    protected fun reloadSessionMeasurements(session: Session) {
+        DatabaseProvider.runQuery { scope ->
+            val dbSessionWithMeasurements = mSessionsViewModel.reloadSessionWithMeasurements(session.uuid)
+            dbSessionWithMeasurements?.let {
+                val reloadedSession = Session(dbSessionWithMeasurements)
+
+                DatabaseProvider.backToUIThread(scope) {
+                    reloadSession(reloadedSession)
+                }
+            }
+        }
     }
 }
